+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Changelog
 Unreleased
 ----------
 ### Changed
-* Removed the environment variable for port from static and static-serve.
+* Removed the environment variable for port from static and static-serve
 
 4.3.0 - (July 17, 2018)
 ----------